Fixed Migrations

+++ database/migrations/2025_09_25_120100_create_countries_table.php
@@ -10,7 +10,7 @@ return new class extends Migration
     {
         Schema::create('countries', function (Blueprint $table) {
             $table->id();
-            $table->unsignedSmallInteger('capital_city')->nullable();
+            $table->unsignedBigInteger('capital_city')->nullable();
             $table->string('name', 255);
             $table->string('name_en', 255);
         });
